Refractor code: dividing the form into 2 pages form
